pyproject.toml
[tool.poetry]
name = "ytransport"
version = "0.1.0"
description = ""
authors = ["Janusz Janowski <januszjanowski@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.9"
tortoise-orm = "^0.18.1"
aerich = "^0.6.2"
aiohttp = "^3.8.1"
asyncpg = "^0.25.0"
aiohttp-jinja2 = "^1.5"
environs = "^9.4.0"
aiohttp-session = {extras = ["secure"], version = "^2.10.0"}
PyNaCl = "^1.5.0"

[tool.poetry.dev-dependencies]
pytest = "^6.1"
black = "^21.12b0"
isort = "^5.10.1"
flake8 = "^4.0.1"
pytest-asyncio = "^0.17.2"

[tool.aerich]
tortoise_orm = "ytransport.conf.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[tool.black]
line_length = 127

[tool.isort]
multi_line_output = 3
line_length = 127
include_trailing_comma = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]
asyncio_mode = "strict"